requirements.txt
# GUI and functionality
tk
msal
paramiko

# Azure and Terraform integration
azure-identity
azure-mgmt-resource
azure-mgmt-network
azure-mgmt-compute
